Moved files to ontolearn directory
